C#
Average Student Grades

+++ C# Advanced/C# Advanced - Exercises/Sets and Dictionaries Advanced - Lab/Average Student Grades/Average Student Grades/Program.cs
@@ -3,25 +3,25 @@
     public static void Main(string[] args)
     {
         int n = int.Parse(Console.ReadLine());
-        Dictionary<string, List<double>> dict = new Dictionary<string, List<double>>();
+        Dictionary<string, List<decimal>> dict = new Dictionary<string, List<decimal>>();
         for (int i = 0; i < n; i++)
         {
             string[] split = Console.ReadLine().Split().ToArray();
             if (dict.ContainsKey(split[0]))
             {
-                dict[split[0]].Add(double.Parse(split[1]));
+                dict[split[0]].Add(decimal.Parse(split[1]));
             }
             else
             {
-                List<double> list = new List<double>();
-                list.Add(double.Parse(split[1]));
+                List<decimal> list = new List<decimal>();
+                list.Add(decimal.Parse(split[1]));
                 dict.Add(split[0], list);
             }
         }
         foreach (var item in dict)
         {
             string name = item.Key;
-            List<double> list = item.Value;
+            List<decimal> list = item.Value;
             Console.Write($"{name} -> ");
             for (int i = 0; i < list.Count; i++)
             {
